(tae): update stl 10
