+++ pkg/vm/engine/tae/stl/container/vector.go
@@ -1,0 +1,14 @@
+package container
+
+func New[T any](opts ...*Options) *Vector[T] {
+	var v T
+	_, ok := any(v).([]byte)
+	if !ok {
+		return &Vector[T]{
+			Vector: NewStdVector[T](opts...),
+		}
+	}
+	return &Vector[T]{
+		Vector: NewStrVector[T](opts...),
+	}
+}
